Restructure domain layer.
