+++ src/domain/users/mod.rs
@@ -1,0 +1,5 @@
+pub mod errors;
+pub mod models;
+
+pub use errors::*;
+pub use models::*;
